# src/main/resources/Promotions_Digital_Enablement-api-spec.yaml
openapi: 3.0.3
info:
  title: Promotions Digital Enablement API
  description: Promotions Digital Enablement API Specification
  termsOfService: https://developer.mastercard.com/terms-of-use
  contact:
    name: API Support
    url: 'https://developer.mastercard.com/support'
    email: apisupport@mastercard.com
  version: 1.1.0
servers:
  - url: https://api.mastercard.com/loyalty/rewards
    description: Production API
  - url: https://sandbox.api.mastercard.com/loyalty/rewards
    description: Sandbox API
...
paths:
...
  /promotions:
    get:
      tags:
        - Promotion Management
      summary: Provides clients the ability to retrieve list of promotions based on the request parameters provided.
      description: This endpoint provides clients with the ability to retrieve all active promotions using the company id, program id, and/or account id.
      operationId: getPromotions
      parameters:
        - $ref: '#/components/parameters/RewardsCompanyId'
        - $ref: '#/components/parameters/ProgramId'
        - $ref: '#/components/parameters/PromotionId'
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/PromotionState'
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          $ref: '#/components/responses/ActivePromotion'
        '400':
          $ref: '#/components/responses/PromotionsBadRequest'
        '401':
          $ref: '#/components/responses/PromotionsUnauthorized'
        '403':
          $ref: '#/components/responses/PromotionsForbidden'
        '404':
          $ref: '#/components/responses/PromotionsNotFound'
  /promotions/{promotion_id}/details:
    get:
      tags:
        - Promotion Management
      summary: Provides clients the ability to retrieve promotion details of particular promotion id.
      description: This endpoint provides clients with the ability to retrieve promotion details using the promotion id.
      operationId: getPromotionDetails
      parameters:
        - $ref: '#/components/parameters/PromotionIdPathParam'
      responses:
        '200':
          $ref: '#/components/responses/PromotionDetail'
        '400':
          $ref: '#/components/responses/PromotionsBadRequest'
        '401':
          $ref: '#/components/responses/PromotionsUnauthorized'
        '403':
          $ref: '#/components/responses/PromotionsForbidden'
        '404':
          $ref: '#/components/responses/PromotionsNotFound'
...
  /transactions:
    get:
      tags:
        - Transactions
      summary: Provides clients the ability to retrieve transaction data.
      description: This endpoint provides clients with the ability to retrieve transaction data related to a specific promotion.
      operationId: getTransactions
      parameters:
        - $ref: '#/components/parameters/AccountIdParam'
        - $ref: '#/components/parameters/FromDateParam'
        - $ref: '#/components/parameters/ToDateParam'
        - $ref: '#/components/parameters/PromotionIdParam'
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          $ref: '#/components/responses/ResponseOkTransaction'
        '400':
          $ref: '#/components/responses/ResponseBadRequest'
        '401':
          $ref: '#/components/responses/ResponseUnauthorizedError'
        '403':
          $ref: '#/components/responses/ResponseForbidden'
        '404':
          $ref: '#/components/responses/ResponseNotFound'
components:
...
  parameters:
    FromDateParam:
      name: from_date
      in: query
      description: Start date from which the client want to view the transactions
      schema:
        type: string
        example: 2023-01-01
        format: date
    ToDateParam:
      name: to_date
      in: query
      description: End date till which the client want to view the transactions
      schema:
        type: string
        example: 2024-01-01
        format: date
    OffsetParam:
      name: offset
      in: query
      description: The number of items to offset the start of the list from (Pagination)
      schema:
        type: integer
        format: int32
        example: 0
        default:
    LimitParam:
      name: limit
      in: query
      description: Can be used to limit the amount of results returned from a query (Pagination)
      schema:
        type: integer
        format: int32
        example: 15
        default: 25
...
    AccountIdParam:
      name: account_id
      in: query
      description: Unique Account identifier assigned by Mastercard and provided in the Account enrollment response
      required: false
      schema:
        type: string
        format: uuid
      example: "f35e51fe-bc77-432f-b412-3800e3c04e5c"
    PromotionIdParam:
      name: promotion_id
      in: query
      description: The unique identifier of the Promotion created and assigned by Mastercard during promotion creation process
      schema:
        type: string
        format: uuid
      example: "2aee0729-3a1a-451a-affa-84f3a6bc5acc"
...
    RewardsCompanyId:
      name: rewards_company_id
      in: query
      description: Mastercard assigned unique identifier for a client. Created and assigned during company creation process
      example: "383a74b4-38b3-4f2a-a85d-6b2e26ad665"
      schema:
        type: string
        format: uuid
    ProgramId:
      name: program_id
      in: query
      description: Unique Program Identifier for a client promotion, available from promotion setup
      example: "aae5ce99-58d1-47f4-8b6a-43b427be9d74"
      schema:
        type: string
        format: uuid
    PromotionId:
      name: promotion_id
      in: query
      description: Unique Promotion Identifier for a client reward program, available from promotion setup
      example: "de0e1456-e678-4270-b027-fc9950b8aba5"
      schema:
        type: string
        format: uuid
    PromotionIdPathParam:
      name: promotion_id
      in: path
      description: Unique Promotion Identifier for a client reward program, available from promotion setup
      example: "de0e1456-e678-4270-b027-fc9950b8aba5"
      required: true
      schema:
        type: string
        format: uuid
    AccountId:
      name: account_id
      in: query
      description: Unique Account identifier assigned by Mastercard and provided in the Account enrollment response
      example: "f35e51fe-bc77-432f-b412-3800e3c04e5c"
      schema:
        type: string
        format: uuid
    PromotionState:
      name: promotion_state
      in: query
      description: Valid states are
        - INACTIVE - Promotion that has never been activated or already expired
        - ACTIVE - Active promotion that is live in a given environment.
        - VALIDATE - Promotion that is not active and is in test state
        - INVALID - promotion is an invalid state
      example: "ACTIVE"
      schema:
        type: string
        default: all
  responses:
...
    PromotionsBadRequest:
      description: Bad Request with errors for the promotions endpoints
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsList'
    PromotionsUnauthorized:
      description: Unauthorized response with errors for the promotions endpoints
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsList'
    PromotionsForbidden:
      description: Forbidden response with errors for the promotions endpoints
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsList'
          examples:
            FORBIDDEN:
              $ref: '#/components/examples/Forbidden'
    PromotionsNotFound:
      description: Not Found with errors for the promotions endpoints
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsList'
    ResponseOkTransaction:
      description: OK response including paged transaction data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PagedTransaction'
    ResponseBadRequest:
      description: Bad Request with errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsList'
    ResponseUnauthorizedError:
      description: Unauthorized error with errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsList'
...
    ResponseForbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsList'
          examples:
            FORBIDDEN:
              $ref: '#/components/examples/Forbidden'
    ResponseNotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsList'
    ActivePromotion:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Promotions"
    PromotionDetail:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PromotionDetail"
...
  schemas:
...
    ErrorItem:
      type: object
      properties:
        Description:
          type: string
          description: Describes Error Type
          example: "Invalid account_id received"
        Details:
          type: string
          description: Describes details of Error Type
          example: "null"
        ReasonCode:
          type: string
          description: ReasonCode for Error Type
          example: "INVALID_INPUT"
        Recoverable:
          type: boolean
          description: Indicates if the Error Type is Recoverable
          example: false
        Source:
          type: string
          description: Indicates Source of Error Type
          example: "VALIDATION_ERROR"
    Errors:
      required:
        - Error
      type: object
      properties:
        Error:
          type: array
          description: List of ErrorItems
          items:
            $ref: '#/components/schemas/ErrorItem'
    ErrorsList:
      required:
        - Errors
      type: object
      properties:
        Errors:
          $ref: '#/components/schemas/Errors'
    PagedTransaction:
      type: object
      properties:
        count:
          type: integer
          description: The number of items that are in this offset batch
          format: int32
          example: 1
        offset:
          type: integer
          description: The number of items to offset the start of the list from
          format: int32
          example: 0
        limit:
          type: integer
          description: The number of items in the list to be limited to
          format: int32
          example: 25
        total:
          type: integer
          description: The total number of items that are in the entire collection
          format: int64
          example: 1
        items:
          type: array
          description: Items in transaction response
          items:
            $ref: '#/components/schemas/GetTransaction'
...
    Account:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
          description: Unique Account identifier assigned by Mastercard and provided in the Account enrollment response.
          example: "f35e51fe-bc77-432f-b412-3800e3c04e5c"
        activation:
          type: boolean
          description: Specifies if the account is opted-in or out for this promotion.
          example: true
        activationDateTime:
          type: string
          format: date-time
          description: Specifies the date when user has optedin or opted out of a promotion.
          example: "2020-06-04T15:19:31.827Z"
    ExternalTargetRecord:
      type: object
      properties:
        externalTargetCode:
          type: string
          description: Code for the associated external target
          example: "pYtYI3Otdn"
        externalBeginDateTime:
          type: string
          format: date-time
          description: Begin date of external target record
          example: "2020-06-17T20:19:55.763Z"
        beginDateTime:
          type: string
          format: date-time
          description: Begin date of promotion association to external target
          example: "2020-05-18T20:19:55.763Z"
        endDateTime:
          type: string
          format: date-time
          description: End date of promotion association to external target
          example: "2020-07-18T20:19:55.765Z"
        entityType:
          type: string
          description: Indicates at which level the external target is associated - either 'A' for Account or 'H' for Household
          example: "A"
        entityId:
          type: string
          format: uuid
          description: Code for the associated external target
          example: "7e747d4a-b585-4990-8756-53996b222804"
        source:
          type: string
          description: Unique entity identifier for external target type which can be the primary account id or just an associated account id
          example: "GRP"
        statusCode:
          type: string
          description: Status code for the associated external target - 'Q', 'P', 'D'
          example: "Q"
        eventDateTime:
          type: string
          format: date-time
          description: Event date time of record sync with the specific source - 'MRS' or 'GRP'
          example: "2020-07-18T20:19:55.765Z"
        externalReferenceId:
          type: string
          description: Identifier for company the external target is associated with
          example: "90909090"
    Promotion:
      type: object
      properties:
        accounts:
          type: array
          description: List of Accounts
          items:
            $ref: "#/components/schemas/Account"
        externalTargetRecords:
          type: array
          description: External Target Records
          items:
            $ref: "#/components/schemas/ExternalTargetRecord"
        activationRequired:
          type: boolean
          example: true
          description: |
            Specifies if a promotion is enabled for Opt-In/Opt-Out.
            - true = Opt-in/out enabled
            - false = Opt-in/out not enabled, account is opted into promotion by default if targeted.
        targetingEnabled:
          type: boolean
          example: false
          description: |
            Specifies if a promotion is enabled for targeting.
            - true = Targeting enabled, account has to be targeted for transactions to qualify.
            - false = Targeting not enabled, all accounts are eligible into promotion by default.
        beginDateTime:
          type: string
          format: date-time
          description: Promotion start date.
          example: "2020-05-05T15:19:31.823Z"
        description:
          type: string
          description: Promotion description.
          example: "Promotion for digital rewards"
        endDateTime:
          type: string
          format: date-time
          description: Promotion end date.
          example: "2020-07-05T15:19:31.827Z"
        scoringBeginDateTime:
          type: string
          format: date-time
          description: Scoring begin date.
          example: "2023-02-01T15:15:31.827Z"
        scoringEndDateTime:
          type: string
          format: date-time
          description: Scoring end date.
          example: "2023-02-01T15:15:31.827Z"
        postingBeginDateTime:
          type: string
          format: date-time
          description: posting begin date.
          example: "2023-02-01T15:15:31.827Z"
        postingEndDateTime:
          type: string
          format: date-time
          description: posting end date.
          example: "2023-02-01T15:15:31.827Z"
        transactionBeginDateTime:
          type: string
          format: date-time
          description: transaction begin date.
          example: "2023-02-01T15:15:31.827Z"
        transactionEndDateTime:
          type: string
          format: date-time
          description: transaction end date.
          example: "2023-02-01T15:15:31.827Z"
        frequency:
          type: string
          description: Promotion frequency.
          example: "RECURRING"
        name:
          type: string
          description: Promotion name.
          example: "Digital Rewards"
        programId:
          type: string
          format: uuid
          description: Unique program identifier.
          example: "aae5ce99-58d1-47f4-8b6a-43b427be9d74"
        promotionId:
          type: string
          format: uuid
          description: Unique promotion identifier.
          example: "de0e1456-e678-4270-b027-fc9950b8aba5"
        timeZoneId:
          type: string
          description: Time zone of a promotion.
          example: "America/Chicago"
        statusCode:
          type: string
          description: Promotion status.
          example: "ACTIVE"
        statusChangedDateTime:
          type: string
          format: date-time
          description: The time when the status was updated.
          example: "2020-06-05T15:19:31.827Z"
...
    Promotions:
      type: object
      properties:
        count:
          type: integer
          format: int32
          example: 1
          description: The number of items that are in this offset batch
        offset:
          type: integer
          format: int32
          example: 1
          description: The number of items to offset the start of the list from
        limit:
          type: integer
          format: int32
          example: 25
          description: The number of items in the list to be limited to
        total:
          type: integer
          format: int64
          example: 1
          description: The total number of items that are in the entire collection
        items:
          type: array
          description: List of Promotions
          items:
            $ref: "#/components/schemas/Promotion"
    PromotionDetail:
      type: object
      properties:
        caps:
          type: array
          description: List of Caps applied to the Promotion
          items:
            $ref: '#/components/schemas/PromotionCap'
        promotionId:
          type: string
          description: Unique promotion identifier
          example: "de0e1456-e678-4270-b027-fc9950b8aba5"
        rewardConfig:
          type: object
          description: Rewards information
          properties:
            promotionId:
              type: string
              format: uuid
              description: Unique promotion identifier
              example: "de0e1456-e678-4270-b027-fc9950b8aba5"
            referenceId:
              type: string
              format: uuid
              description: Unique reward config identifier
              example: "de89f43c-8fbe-45ae-b5d2-ab965ecaa956"
            rebateText:
              type: string
              description: Specifies the text that shows up on cardholders billing statement for a purchase (Only applicable for a rebate)
              example: "REBATE"
            rebateTextNegative:
              type: string
              description: Specifies the text that shows up on cardholders billing statement for a return (Only applicable for a rebate)
              example: "RETURN"
            rewardType:
              type: string
              description: |
                Specifies the type of reward.
                - CASHBACK - Cashback
                - POINTS - Points
                - REWARD_ITEM - Reward Item
                - BENEFIT - Benefit
              example: "CASHBACK"
            pointType:
              type: string
              description: |
                Specifies how the points are awarded (Only applicable if rewardType is POINTS).
                - BASE_POINTS - Base points
                - BONUS_POINTS - Bonus points
              example: "BASE_POINTS"
            awardType:
              type: string
              description: |
                Specifies if the award is going to be a fixed amount or if its going to be based on transaction amount.
                - FIXED - Fixed amount
                - PERCENTAGE - Percentage of transaction amount
              example: "FIXED"
            applyAwardTo:
              type: string
              description: |
                Specifies if the award is applied to transacting account or the household primary account.
                - TRANSACTING_ACCOUNT - Transacting account
                - PRIMARY_ACCOUNT - Household primary account
              example: "TRANSACTING_ACCOUNT"
            scoreBasedOn:
              type: string
              description: |
                Specifies how the transactions are scored.
                - DE4_TRANSACTION_AMOUNT - Transaction amount
                - DE5_SETTLEMENT_AMOUNT - Settlement amount
                - DE6_CARDHOLDER_BILLING_AMOUNT - Cardholder billing amount
              example: "DE4_TRANSACTION_AMOUNT"
            rewardRate:
              type: string
              description: If the awardType is FIXED, this field specifies the amount that is awarded. And if the awardType is not fixed then it specifies the percentage of transaction amount.
              example: "10"
        thresholds:
          type: array
          description: List of Thresholds applied to the Promotion
          items:
            $ref: '#/components/schemas/PromotionThreshold'
    PromotionThreshold:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique threshold identifier
          example: "fbbbba72-141d-41f5-9ace-69129f971010"
        type:
          type: string
          description: |
            Specifies if the threshold is based on transaction count or transaction spend amount.
            - COUNT - Transaction count
            - SPEND - Transaction spend amount
          example: "AMOUNT"
        fieldType:
          type: string
          description: |
            If the type selected is SPEND, then this field specifies which amount value is used to accumulate toward the threshold.
            - DE04 - Transaction Amount
            - DE06 - Cardholder Billing Amount
          example: "DE04"
        value:
          type: number
          format: double
          description: The spend amount or the count that must be reached before the award is given
          example: 1000.00
        level:
          type: string
          description: |
            Specifies the level at which the threshold is applied.
            - A - Account level (All accounts have their own threshold)
            - H - Household level (Each account contributes toward reaching the threshold)
          example: "A"
        periodInDaysAfterDate:
          type: integer
          format: int32
          description: Number of days the cardholder has to meet the threshold before its expired
          example: 10
        dateFieldType:
          type: string
          description: |
            Specifies the date type used for starting the count on how long the account has for meeting the threshold.
            - ENROLLMENT_DATE - The date on which the cardholder has enrolled to the program
            - TARGET_BEGIN_DATE - The date on which the account is set to be targeted into a promotion
            - ACTIVE_DATE - The date on which enrollment record was added to the system
            - OPTIN_DATE - The date on which the account has opted into the promotion
          example: "ENROLLMENT_DATE"
    PromotionCap:
      type: object
      properties:
        amount:
          type: number
          format: double
          description: Specifies the cap amount
          example: 1000.00
        enabled:
          type: boolean
          description: Indicates if capping is enabled for the promotion
          example: TRUE
        id:
          type: string
          format: uuid
          description: Unique cap identifier
          example: "59266873-45c9-418f-a9dc-1b0f00d0bc16"
        level:
          type: string
          description: |
            Specifies the level at which cap is applied.
            - A - Account level
            - P - Household level
            - T - Transaction level
          example: "A"
        period:
          type: string
          description: |
            Specifies the period for which cap is applied.
            - M - Monthly
            - Y - Yearly
            - L - Lifetime
          example: "M"
...
  examples:
...
    Forbidden:
      value:
        Errors:
          Error:
            - Source: Loyalty-Enrollment
              ReasonCode: FORBIDDEN
              Description: Client does not have permission.
              Recoverable: true
              Details: null